separate mappers files and refactor handling exception
