+++ repository/search/src/main/java/com/repository/search/exception/AflamiException.kt
@@ -8,4 +8,5 @@ open class LocalDataSourceException(message:String):AflamiException(message)
 class NoInternetConnectionException():NetworkConnectionException("Please connect your device to the internet")
 class NoDataForCountryException():LocalDataSourceException("No movies found for the given country")
 class NoDataForActorException():LocalDataSourceException("No media found for the given actor")
-class NoDataWereFoundException():LocalDataSourceException("No media matched your search term")
+class NoDataForSearchException():LocalDataSourceException("No media matched your search term")
+class NoCategoriesFoundException():LocalDataSourceException("No categories found ")
