+++ repository/search/src/main/java/com/repository/search/mapper/SearchHistoryMapper.kt
@@ -1,0 +1,17 @@
+package com.repository.search.mapper
+
+import com.domain.search.model.SearchHistoryModel
+import com.repository.search.entity.SearchHistoryEntity
+
+
+fun List<SearchHistoryEntity>.toSearchHistories(): List<SearchHistoryModel> {
+    return this.map { it.toSearchHistoryModel() }
+}
+
+fun SearchHistoryEntity.toSearchHistoryModel(): SearchHistoryModel {
+    return SearchHistoryModel(
+
+        searchTitle = this.searchQuery,
+        searchDate = this.searchDate.toString()
+    )
+}
