+++ repository/search/src/main/java/com/repository/search/repository/SearchMediaRepositoryImpl.kt
@@ -1,23 +1,19 @@
 package com.repository.search.repository
 
 import com.domain.search.model.Media
-import com.domain.search.model.MediaType
 import com.domain.search.repository.SearchMediaRepository
 import com.repository.search.NetworkConnectionChecker
 import com.repository.search.dataSource.local.HistoryLocalDataSource
 import com.repository.search.dataSource.local.MediaLocalDataSource
 import com.repository.search.dataSource.remote.SearchRemoteDataSource
-import com.repository.search.dto.ResultDto
-import com.repository.search.dto.SearchDto
-import com.repository.search.entity.MediaEntity
-import com.repository.search.entity.MediaTypeEntity
 import com.repository.search.exception.NoDataForActorException
 import com.repository.search.exception.NoDataForCountryException
-import com.repository.search.exception.NoDataWereFoundException
+import com.repository.search.exception.NoDataForSearchException
 import com.repository.search.exception.NoInternetConnectionException
+import com.repository.search.mapper.toMediaEntities
+import com.repository.search.mapper.toMedias
 import com.repository.search.util.getCurrentDate
 import kotlinx.datetime.DateTimeUnit
-import kotlinx.datetime.LocalDate
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.plus
 import kotlinx.datetime.toInstant
@@ -125,67 +121,7 @@ class SearchMediaRepositoryImpl(
             }
             return mediaLocalDataSource.getMediaByTitleQuery(query = query).toMedias()
         } catch (_: Exception) {
-            throw NoDataWereFoundException()
+            throw NoDataForSearchException()
         }
     }
-}
-
-
-fun List<MediaEntity>.toMedias() = this.map { it.toMedia() }
-
-fun MediaEntity.toMedia(): Media {
-    return Media(
-        id = this.id,
-        imageUri = this.imageUri,
-        title = this.title,
-        type = this.type.toMediaType(),
-        categories = this.category,
-        yearOfRelease = this.yearOfRelease,
-        rating = this.rating
-    )
-}
-
-fun MediaTypeEntity.toMediaType(): MediaType {
-    return when (this) {
-        MediaTypeEntity.MOVIE -> MediaType.MOVIE
-        MediaTypeEntity.TVSHOW -> MediaType.TVSHOW
-    }
-}
-
-
-fun ResultDto.toMediaEntity(
-    searchQuery: String,
-    actor: List<String>,
-    country: String
-): MediaEntity? {
-    val title = this.title ?: this.name ?: return null
-    val image = this.posterPath ?: this.profilePath ?: ""
-    val releaseDateStr = this.releaseDate ?: this.firstAirDate ?: return null
-
-    return MediaEntity(
-        searchQuery = searchQuery,
-        imageUri = image,
-        title = title,
-        type = when (this.mediaType) {
-            "movie" -> MediaTypeEntity.MOVIE
-            "tv" -> MediaTypeEntity.TVSHOW
-            else -> return null
-        },
-        category = this.genreIds?.map { it.toString() } ?: emptyList(),
-        yearOfRelease = LocalDate.parse(releaseDateStr),
-        rating = this.voteAverage ?: 0.0,
-        country = country,
-        actor = actor
-    )
-}
-
-
-fun SearchDto.toMediaEntities(
-    query: String,
-    actor: List<String> = emptyList(),
-    country: String = ""
-): List<MediaEntity> {
-    return results?.mapNotNull {
-        it.toMediaEntity(query, actor, country)
-    } ?: emptyList()
 }
